Using a request/response instead of attempting to single page the logout. This is finally working for basic auth so we're checking pointing all this.
Commented out the service logout for single page.  I'll leave it for now if
we can get a good understanding of it.

Added Angular routing for logout... not sure if it matters, makes sense, or
is right.

Added logging to the "/user" service.

Unwired the form login Spring security filter.  We'll try enabling it after
this.

Added a UserContextService interface and implementation.

Added a UserDetailsService implementation.  Added a UserDto which implements
the UserDetails Spring Security interface.

Added minor notes to the README.md for front end builds.  Not settled on how
this is working.  More to come on this soon.