***Bingo***
So this appears to do everything we care about.  It's a combination of Form
and Basic Spring Security.  The form based is working for certain pages
(/home in this case) and Basic is working in a stateless way as expected where
you can send the credentials in the URL: http://user:password@localhost:8080/basic/resource.
So what this means is a different page from the browser will not let the /home
resource be accessed without authentication but the basic resource with creds
is available all day long.  This is super good.